+++ mixpanel-dispatcher/src/main/java/com/batch/android/dispatcher/mixpanel/MixpanelDispatcher.java
@@ -43,6 +43,11 @@ public class MixpanelDispatcher implements BatchEventDispatcher
     private static final String BATCH_TRACKING_ID = "batch_tracking_id";
 
     /**
+     * Key used to dispatch the webview click analytics ID
+     */
+    private static final String BATCH_WEBVIEW_ANALYTICS_ID = "batch_webview_analytics_id";
+
+    /**
      * Event name used when logging on Mixpanel
      */
     private static final String NOTIFICATION_DISPLAY_NAME = "batch_notification_display";
@@ -51,7 +56,9 @@ public class MixpanelDispatcher implements BatchEventDispatcher
     private static final String MESSAGING_SHOW_NAME = "batch_in_app_show";
     private static final String MESSAGING_CLOSE_NAME = "batch_in_app_close";
     private static final String MESSAGING_AUTO_CLOSE_NAME = "batch_in_app_auto_close";
+    private static final String MESSAGING_CLOSE_ERROR_NAME = "batch_in_app_close_error";
     private static final String MESSAGING_CLICK_NAME = "batch_in_app_click";
+    private static final String MESSAGING_WEBVIEW_CLICK_NAME = "batch_in_app_webview_click";
     private static final String UNKNOWN_EVENT_NAME = "batch_unknown";
 
     MixpanelAPI mixpanelInstance = null;
@@ -111,6 +118,11 @@ public class MixpanelDispatcher implements BatchEventDispatcher
         mixpanelParams.put(CAMPAIGN, payload.getTrackingId());
         mixpanelParams.put(MEDIUM, "in-app");
         mixpanelParams.put(BATCH_TRACKING_ID, payload.getTrackingId());
+
+        String webViewAnalyticsId = payload.getWebViewAnalyticsID();
+        if (webViewAnalyticsId != null) {
+            mixpanelParams.put(BATCH_WEBVIEW_ANALYTICS_ID, webViewAnalyticsId);
+        }
 
         String deeplink = payload.getDeeplink();
         if (deeplink != null) {
@@ -229,8 +241,12 @@ public class MixpanelDispatcher implements BatchEventDispatcher
                 return MESSAGING_CLOSE_NAME;
             case MESSAGING_AUTO_CLOSE:
                 return MESSAGING_AUTO_CLOSE_NAME;
+            case MESSAGING_CLOSE_ERROR:
+                return MESSAGING_CLOSE_ERROR_NAME;
             case MESSAGING_CLICK:
                 return MESSAGING_CLICK_NAME;
+            case MESSAGING_WEBVIEW_CLICK:
+                return MESSAGING_WEBVIEW_CLICK_NAME;
         }
         return UNKNOWN_EVENT_NAME;
     }